requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.2
django-discover-runner==1.0
django-environ==0.4.5
django-localflavor==3.0.1
django-model-utils==4.1.1
django-role-permissions==3.1.1
environ==1.0
idna==2.10
mercadopago==2.0.6
Pillow==8.2.0
python-stdnum==1.16
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
